% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Energy_realizations}
\alias{Energy_realizations}
\title{Simulation of Energy generation}
\usage{
Energy_realizations(Head_largerealizations, Outflows, efficieny, t, ntimes)
}
\arguments{
\item{Head_largerealizations}{A large matrix of \code{2*ntimes+1} simulations/realizations for Head. Each rows
represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Head data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.}

\item{Outflows}{The is the matrix tranformation of outflows data (decision variables) generated by SQP algorithm in a vector form.
where each column is the day for \code{t} days optimization period. The rows represents the outflows of GCL,
LWG and MCN respectively.}

\item{t}{An integer indicating the number of timesteps. If we consider daily timestep for 14 day optimization
period, \code{t} = 14.}

\item{ntimes}{An integer indicating the number of simulations/realizations.}

\item{efficiency}{This is considered as 0.75.}
}
\value{
A large matrix of \code{2*ntimes+1} simulations/realizations for Energy generation. Each rows
represent a single realization, \code{2*ntimes+1} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Energy generation data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
}
\description{
Generating total \code{2*ntimes+1} Energy generation simulations for each of \code{t} days optimization period
}
\details{
Generating total \code{2*ntimes+1} Energy generation simulations for each of \code{t} days optimization period.
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Energy generation data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
This function is required to quantify the unertainity on Energy generation of Grand Coulle and Lower Granite.
Also, the result from this function is required to simulate Revenue.
}
\examples{
Head_largerealizations <- Head_largerealizations
Outflows <- Outflows_matrixform
efficiency <- 0.75
t <- 14
ntimes <-100
}
