% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Forebay_realizations}
\alias{Forebay_realizations}
\title{Simulation of Forebay Elevation}
\usage{
Forebay_realizations(Storage_largerealizations, Fb_coeff, t, ntimes)
}
\arguments{
\item{Storage_largerealizations}{A large matrix of \code{2*ntimes+1} simulations/realizations for Storage. Each rows
represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Storage data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.}

\item{Fb_coeff}{This is coefficients to calculate the Forebay. The coefficients are provided as input data and the values
has been obtained separately by the regression model. Each row is the coefficient data for GCL, LWG and MCN respectively.}

\item{t}{An integer indicating the number of timesteps. If we consider daily timestep for 14 day optimization
period, \code{t} = 14.}

\item{ntimes}{An integer indicating the number of simulations/realizations.}
}
\value{
A large matrix of \code{2*ntimes+1} simulations/realizations for Forebay. Each rows
represent a single realization, \code{2*ntimes+1} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Forebay data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
}
\description{
Generating total \code{2*ntimes+1} Forebay simulations for each of \code{t} days optimization period
}
\details{
Generating total \code{2*ntimes+1} Forebay simulations for each of \code{t} days optimization period.
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Forebay data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
This function is required to quantify the unertainity on Forebay of Grand Coulle and Lower Granite.
Also, the result from this function is required to simulate Tailwater and Head.
}
\examples{
Storage_largerealizations <- Storage_largerealizations
Fb_coeff <- Fb_coeff
t <- 14
ntimes <-100
Forebay_realizations(Storage_largerealizations, Fb_coeff, t, ntimes)
}
