% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{get_samples}
\alias{get_samples}
\title{Generating random samples}
\usage{
get_samples(t, ntimes, GCLInflowsdata, LWGInflowsdata, Pricedata)
}
\arguments{
\item{t}{An integer indicating the number of timesteps. If we consider daily timestep for 14 day optimization
period, \code{t} = 14.}

\item{ntimes}{An integer indicating the number of simulations/realizations.}

\item{GCLInflowsdata}{A matrix of inputs of predicted Inflows of GCL where each row (m) is
a day of optimization period \code{t} and each column (n) represent a single realization. The first column constains
the characters mentioning the days of the inflows.}

\item{LWGInflowsdata}{A matrix of inputs of predicted Inflows of LWG where each row (m) is
a day of optimization period \code{t} and each column (n) represent a single realization. The first column constains
the characters mentioning the days of the inflows.}

\item{Pricedata}{A matrix of inputs of predicted Price where each row (m) represent a single realization
and each column (n) is a day of optimization period \code{t}. The first column constains
the characters mentioning the realization number.}
}
\value{
A large list of \code{ntimes} random normal samples of Inflows and \code{ntimes^2} random uniform samples of Price for \code{t} = 14 days.
For example, in the samples of Inflows to Grand Coulle reservoir (GCLInflowsdata_samples)
inside the list, each rows represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days
}
\description{
Generating random samples using Monte Carlo approach
}
\details{
Generating \code{ntimes} normal random samples of inflows of Grand Coulee and Lower Granite for each day of 
of the optimization period, based on the means and standard deviations of 27 predicted realizations 
of historic inflows data. Also generates \code{ntimes^2} uniform random samples of Price for each day of the
optimization period, based on the means and standard deviations of 5 predicted realizations.
}
\examples{
t <- 14
ntimes <-500
GCLInflowsdata <- GCLInflowsdata
LWGInflowsdata <- LWGInflowsdata
Pricedata<-Pricedata
get_samples(t, ntimes,GCLInflowsdata,LWGInflowsdata,Pricedata)
}
