% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Head_realizations}
\alias{Head_realizations}
\title{Simulation of Head}
\usage{
Head_realizations(Forebay_largerealizations, Tailwater_largerealizations)
}
\arguments{
\item{Forebay_largerealizations}{A large matrix of \code{2*ntimes+1} simulations/realizations for Forebay. Each rows
represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Forebay data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.}

\item{Tailwater_largerealizations}{A large matrix of \code{2*ntimes+1} simulations/realizations for Tailwater. Each rows
represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Tailwater data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.}
}
\value{
A large matrix of \code{2*ntimes+1} simulations/realizations for Head. Each rows
represent a single realization, \code{2*ntimes+1} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Head data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
}
\description{
Generating total \code{2*ntimes+1} Head simulations for each of \code{t} days optimization period
}
\details{
Generating total \code{2*ntimes+1} Head simulations for each of \code{t} days optimization period.
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Head data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
This function is required to quantify the unertainity on Head of Grand Coulle and Lower Granite.
Also, the result from this function is required to simulate Energy generation.
}
\examples{
Forebay_largerealizations <- Forebay_largerealizations
Tailwater_largerealizations <- Tailwater_largerealizations
Head_realizations(Forebay_largerealizations, Tailwater_largerealizations)
}
