% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{mean_sd_sim}
\alias{mean_sd_sim}
\title{Quantifying means and standard deviations using Monte Carlo approach.}
\usage{
mean_sd_sim(samples, Storage_initial, Inflow_initial, Outflow_initial,
  Forebay_initial, Tailwater_initial, Outflows, Fb_coeff, Tw_coeff,
  delta_t, efficieny, t, r, n_samples)
}
\arguments{
\item{samples}{A large list of \code{ntimes} random normal samples of Inflows and \code{ntimes^2} random uniform samples of Price for \code{t} = 14 days.
For example, in the samples of Inflows to Grand Coulle reservoir (GCLInflowsdata_samples)
inside the list, each rows represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days}

\item{Storage_initial}{A vector of length 3 with initial/current storage of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Inflow_initial}{A vector of length 3 with initial/current inflows of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Outflow_initial}{A vector of length 3 with initial/current outflows of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Forebay_initial}{A vector of length 3 with initial/current forebay of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Tailwater_initial}{A vector of length 3 with initial/current tailwater of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Outflows}{The is the matrix tranformation of outflows data (decision variables) generated by SQP algorithm in a vector form.
where each column is the day for \code{t} days optimization period. The rows represents the outflows of GCL,
LWG and MCN respectively.}

\item{Fb_coeff}{This is coefficients to calculate the Forebay. The coefficients are provided as input data and the values
has been obtained separately by the regression model. Each row is the coefficient data for GCL, LWG and MCN respectively.}

\item{Tw_coeff}{This is coefficients to calculate the Tailwater. The coefficients are provided as input data and the values
has been obtained separately by the regression model. Each column is the coefficient data for GCL, LWG and MCN respectively.}

\item{delta_t}{This is value of time-step in day. In our problem, we have considered daily timesteps for optimization.
Therefore, \code{delta_t} =1}

\item{t}{An integer indicating the number of timesteps. If we consider daily timestep for 14 day optimization
period, \code{t} = 14.}

\item{r}{An integer indicating the number of reservoirs. In this case study, \code{r}=3}

\item{n_samples}{An integer indicating the number of simulations/realizations.}

\item{efficiency}{This is considered as 0.75.}
}
\value{
A large list of simulations along with the mean and standard deviation
for Storage, Forebay, Energy etc. For example, in the simulation matrix of Energy (Energy_realizations) inside the list, each rows
represent a single realization, \code{2*ntimes+1} and each columns represent the time steps,\code{t} in days.
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Energy generation data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
}
\description{
Quantifying means and standard deviations for Storage, Forebay, Tailwater, Head and Energy generation 
using Monte Carlo approach.
}
\details{
Quantifying means and standard deviations for Storage, Forebay, Tailwater, Head and Energy generation 
using Monte Carlo approach. This function is required to evaluate robust objective function and validate
probabilistic constraints
}
\examples{
samples <- samples
Storage_initial <- Current_Storage
Inflow_initial <- Current_Inflows
Outflow_initial <- Current_Outflows
Forebay_initial <- Current_Forebay
Tailwater_initial <- Current_Tailwater
Outflows <- Outflows_matrixform
Fb_coeff <- Fb_coeff
Tw_coeff <- Tw_coeff
delta_t <- 1
efficieny <- 0.75
t <- 14
r<- 3
n_samples <-500
mean_sd_sim(samples, Storage_initial, Inflow_initial, Outflow_initial,  Forebay_initial, Tailwater_initial, Outflows, Fb_coeff, Tw_coeff, delta_t, efficieny, t, r, n_samples)
}
