% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objfuncrevenue.R
\name{obj_Func_eval}
\alias{obj_Func_eval}
\title{Evaluation of Robust Objective Function}
\usage{
obj_Func_eval(X, samples, initial_cond)
}
\arguments{
\item{X}{The is the vector of outflows data (decision variables) generated by SQP algorithm. It is a vector 
of length \code{t}*number of reservoirs. For the case of \code{t} =14 and 3 reservoir system, the length of 
\code{X} = 14*3 = 42}

\item{samples}{A large list of \code{ntimes} random normal samples of Inflows and \code{ntimes^2} random uniform samples of Price for \code{t} = 14 days.
For example, in the samples of Inflows to Grand Coulle reservoir (GCLInflowsdata_samples)
inside the list, each rows represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days}

\item{initial_cond}{A large list for all the initial conditions required 
for the functions \code{mean_sd_sim()}, \code{mean_sd_sim_antithetic()} and \code{Total_Revenue_cal()} 
which are called inside this function}
}
\value{
An integer value of Robust objective function
}
\description{
Calculating the robust objective function value for \code{t} days optimization period
}
\details{
Quantifies the robust objective considering the uncertainities in inflows and Prices of electricity.
To calculate the robust objective, the function calls two other functions:Firstly, \code{mean_sd_sim()} or \code{mean_sd_sim_antithetic()}
based on the user choice to go for MC or Antithetic Variable approach to generate simulations of energy.
After that, another function \code{Total_Revenue_cal()} is called which provides mean and standard deviation of the total revenue for 
\code{t} days optimization period.
}
\examples{
X <- Outflows
samples <- samples
initial_cond <- initial_cond
obj_Func_eval(X, samples, initial_cond)
}
