% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Storage_realizations}
\alias{Storage_realizations}
\title{Simulation of Storage}
\usage{
Storage_realizations(Inflow1, Inflow2, Storage_initial, Inflow_initial,
  Outflow_initial, Outflows, delta_t, t, ntimes)
}
\arguments{
\item{Inflow1}{A large matrix of \code{ntimes} simulations of Inflows of GCL for \code{t} = 14 days. Each rows
represent a single realization and each columns represent the time steps,\code{t} in days}

\item{Inflow2}{A large matrix of \code{ntimes} simulations of Inflows of LWG for \code{t} = 14 days. Each rows
represent a single realization and each columns represent the time steps,\code{t} in days}

\item{Storage_initial}{A vector of length 3 with initial/current storage of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Inflow_initial}{A vector of length 3 with initial/current inflows of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Outflow_initial}{A vector of length 3 with initial/current outflows of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Outflows}{The is the matrix tranformation of outflows data (decision variables) generated by SQP algorithm in a vector form.
where each column is the day for \code{t} days optimization period. The rows represents the outflows of GCL,
LWG and MCN respectively.
\code{Outflows} = 14*3 = 42}

\item{delta_t}{This is value of time-step in day. In our problem, we have considered daily timesteps for optimization.
Therefore, \code{delta_t} =1}

\item{t}{An integer indicating the number of timesteps. If we consider daily timestep for 14 day optimization
period, \code{t} = 14.}

\item{ntimes}{An integer indicating the number of simulations/realizations.}
}
\value{
A large matrix of \code{2*ntimes+1} simulations/realizations for Storage. Each rows
represent a single realization, \code{2*ntimes+1} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Storage data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
}
\description{
Generating total \code{2*ntimes+1} storage simulations for each of \code{t} days optimization period
}
\details{
Generating total \code{2*ntimes+1} storage simulations for each of \code{t} days optimization period.
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Storage data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
This function is required to quantify the unertainity on Storage of Grand Coulle and Lower Granite.
Also, the result from this function is required to simulate forebay elevation.
}
\examples{
Inflow1 <- GCLInflowsdata_samples
Inflow2 <- LWGInflowsdata_samples
Storage_initial <- Current_Storage
Inflow_initial <- Current_Inflows
Outflow_initial <- Current_Outflows
Outflows <- Outflows_matrixform
delta_t <- 1
t <- 14
ntimes <-100
Storage_realizations(Inflow1, Inflow2, Storage_initial, Inflow_initial, Outflow_initial, Outflows, delta_t, t, ntimes)
}
