% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Tailwater_realizations}
\alias{Tailwater_realizations}
\title{Simulation of Tailwater Elevation}
\usage{
Tailwater_realizations(Forebay_largerealizations, Outflow_initial,
  Forebay_initial, Tailwater_initial, Tw_coeff, Outflows, t, ntimes)
}
\arguments{
\item{Forebay_largerealizations}{A large matrix of \code{2*ntimes+1} simulations/realizations for Forebay. Each rows
represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Tailwater data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.}

\item{Outflow_initial}{A vector of length 3 with initial/current outflows of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Forebay_initial}{A vector of length 3 with initial/current forebay of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Tailwater_initial}{A vector of length 3 with initial/current tailwater of Grand Coulee, Lower Granite and McNaire.
for all \code{t} days. This is provided in the model as initial condition of the optimization.}

\item{Tw_coeff}{This is coefficients to calculate the Tailwater. The coefficients are provided as input data and the values
has been obtained separately by the regression model. Each column is the coefficient data for GCL, LWG and MCN respectively.}

\item{Outflows}{The is the matrix tranformation of outflows data (decision variables) generated by SQP algorithm in a vector form.
where each column is the day for \code{t} days optimization period. The rows represents the outflows of GCL,
LWG and MCN respectively.}

\item{t}{An integer indicating the number of timesteps. If we consider daily timestep for 14 day optimization
period, \code{t} = 14.}

\item{ntimes}{An integer indicating the number of simulations/realizations.}
}
\value{
A large matrix of \code{2*ntimes+1} simulations/realizations for Tailwater. Each rows
represent a single realization, \code{2*ntimes+1} and each columns represent the time steps,\code{t} in days   
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Tailwater data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
}
\description{
Generating total \code{2*ntimes+1} Tailwater simulations for each of \code{t} days optimization period
}
\details{
Generating total \code{2*ntimes+1} Tailwater simulations for each of \code{t} days optimization period.
The first \code{ntimes} rows are the simulations for GrandCoulee, next \code{ntimes} rows are the simulations for Lower Granite
and the last row of the matrix is the deterministic Tailwater data for McNaire. McNaire Inflows doesnot have any uncertainity
since it is the total outflows released from GCL and LWG which are deterministic. In future, we can introduce other source
of uncertainity in the inflows of McNaire.
This function is required to quantify the unertainity on Tailwater of Grand Coulle and Lower Granite.
Also, the result from this function is required to simulate Head.
}
\examples{
Forebay_largerealizations <- Forebay_largerealizations
Outflow_initial <- Current_Outflows
Forebay_initial <- Current_Forebay
Tailwater_initial <- Current_Tailwater
Tw_coeff <- Tw_coeff
Outflows <- Outflows_matrixform
t <- 14
ntimes <-100
Tailwater_realizations(Forebay_largerealizations, Outflow_initial, Forebay_initial, Tailwater_initial, Tw_coeff, Outflows, t, ntimes)
}
