% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objfuncrevenue.R
\name{Total_Revenue_cal}
\alias{Total_Revenue_cal}
\title{Quantifying the Total Revenue}
\usage{
Total_Revenue_cal(mean_sd_simulations, Price, t, ntimes)
}
\arguments{
\item{mean_sd_simulations}{A large list of simulations along with the mean and standard deviation
for Storage, Forebay, Energy etc. For example, in the simulation matrix of Energy (Energy_realizations) inside the list, each rows
represent a single realization, \code{ntimes} and each columns represent the time steps,\code{t} in days}

\item{Price}{A large matrix of \code{ntimes^2} simulations of Price for \code{t} = 14 days. Each rows
represent a single realization and each columns represent the time steps,\code{t} in days}

\item{t}{An integer indicating the number of timesteps. If we consider daily timestep for 14 day optimization
period, \code{t} = 14.}

\item{ntimes}{An integer indicating the number of simulations/realizations.}
}
\value{
A vector of length = 2 with the mean and standard deviation of the total revenue for 
\code{t} days optimization period like \code{c(Total_Revenue_Mean, Total_Revenue_sd)}
}
\description{
Quantifying the Total Revenue for \code{t} days optimization period
}
\details{
Quantifies the total revenue generated for \code{t} days considering the uncertainities in inflows and Prices of electricity.
This function is required to evaluate the objective function in the model. For example, if \code{ntimes} = 100 and 
\code{t} = 14, due to 2 sources of inflows uncertaitinity from Grand Coulee and Lower Granite reservoirs, the total 
number of simulations will be \code{ntimes^2} = 10000
}
\examples{
mean_sd_simulations <- mean_sd_simulations
Price <- Price_samples
t <- 14
ntimes <-100
Total_Revenue_cal(mean_sd_simulations, Price, t, ntimes)
}
